[React Redux] Move out Counters reducers to new file
